Fixes the leaderboard bug
