+++ lego-webapp/pages/achievements/leaderboard/event-count/+Page.tsx
@@ -1,0 +1,5 @@
+import LeaderboardTable from '../LeaderboardTable';
+
+const EventCountLeaderboard = () => <LeaderboardTable type="event_count" />;
+
+export default EventCountLeaderboard;
