+++ lego-webapp/pages/achievements/utils.tsx
@@ -3,7 +3,9 @@ import { NavigationTab } from '~/components/NavigationTab/NavigationTab';
 export const AchievementTabs = () => (
   <>
     <NavigationTab href="/achievements">Oversikt</NavigationTab>
-    <NavigationTab href="/achievements/leaderboard">Topplister</NavigationTab>
+    <NavigationTab href="/achievements/leaderboard" matchSubpages>
+      Topplister
+    </NavigationTab>
   </>
 );
