+++ lego-webapp/redux/actions/AchievementActions.ts
@@ -1,7 +1,6 @@
 import { Achievement, User } from '~/redux/actionTypes';
 import { userSchema } from '~/redux/schemas';
 import callAPI from './callAPI';
-import type { APIPromiseResult } from './callAPI';
 import type { ParsedQs } from 'qs';
 import type { PublicUser } from '~/redux/models/User';
 
@@ -11,33 +10,22 @@ export function fetchLeaderboardUsers({
 }: {
   next: boolean;
   query: ParsedQs;
-}): APIPromiseResult<PublicUser[]> {
-  return (dispatch, getState) => {
-    // Claim the next requestId so a response that resolves after a newer
-    // leaderboard fetch has already started (e.g. rapidly switching rank
-    // type) can be recognized as stale and discarded, see users.ts.
-    const requestId = getState().users.leaderboardRequestId + 1;
-
-    return dispatch(
-      callAPI<PublicUser[]>({
-        types: User.FETCH_LEADERBOARD,
-        endpoint: `/achievements/leaderboard/`,
-        query,
-        pagination: {
-          fetchNext: next,
-        },
-        schema: [userSchema],
-        method: 'GET',
-        meta: {
-          errorMessage: 'Henting av brukere feilet',
-          requestId,
-        },
-        discardResultIf: () =>
-          getState().users.leaderboardRequestId !== requestId,
-      }),
-    );
-  };
+}) {
+  return callAPI<PublicUser[]>({
+    types: User.FETCH_LEADERBOARD,
+    endpoint: `/achievements/leaderboard/`,
+    query,
+    pagination: {
+      fetchNext: next,
+    },
+    schema: [userSchema],
+    method: 'GET',
+    meta: {
+      errorMessage: 'Henting av brukere feilet',
+    },
+  });
 }
+
 export function postGettingWood() {
   return callAPI({
     endpoint: `/achievements/getting_wood/`,
